+++ frontend/src/features/consultants/components/skills/education-form.tsx
@@ -4,6 +4,7 @@ import { Input } from "../../../../components/ui/input";
 import { Button } from "../../../../components/ui/button";
 import EducationTable from "./consultant-education-table";
 import type { Education } from "./consultant-education-table";
+import { formatDateInput, parseDate, validateDateRange } from "../../utils/date.utils";
 
 export default function EducationForm() {
     const [educationList, setEducationList] = useState<Education[]>(() => {
@@ -19,24 +20,6 @@ export default function EducationForm() {
     const [startDate, setStartDate] = useState(() => sessionStorage.getItem("education_startDate") || "");
     const [endDate, setEndDate] = useState(() => sessionStorage.getItem("education_endDate") || "");
     const [dateError, setDateError] = useState("");
-
-    const formatDateInput = (value: string) => {
-        const v = value.replace(/\D/g, "");
-        if (v.length >= 5) {
-            return `${v.slice(0, 2)}/${v.slice(2, 4)}/${v.slice(4, 8)}`;
-        } else if (v.length >= 3) {
-            return `${v.slice(0, 2)}/${v.slice(2)}`;
-        }
-        return v;
-    };
-
-    const parseDate = (dateStr: string) => {
-        const parts = dateStr.split("/");
-        if (parts.length === 3 && parts[2].length === 4) {
-            return new Date(Number(parts[2]), Number(parts[1]) - 1, Number(parts[0]));
-        }
-        return null;
-    };
 
     const handleStartDateChange = (e: React.ChangeEvent<HTMLInputElement>) => {
         const newStart = formatDateInput(e.target.value);
@@ -60,21 +43,13 @@ export default function EducationForm() {
     const handleAddEducation = () => {
         if (!institutionName.trim() || !qualification.trim() || !endDate) return;
 
-        const parsedStart = parseDate(startDate);
+        const validationError = validateDateRange(startDate, endDate);
+        if (validationError) {
+            setDateError(validationError);
+            return;
+        }
+
         const parsedEnd = parseDate(endDate);
-        
-        if (startDate && !parsedStart) {
-            setDateError("Please enter a valid Start Date (DD/MM/YYYY)");
-            return;
-        }
-        if (endDate && !parsedEnd) {
-            setDateError("Please enter a valid End Date (DD/MM/YYYY)");
-            return;
-        }
-        if (parsedStart && parsedEnd && parsedEnd < parsedStart) {
-            setDateError("End date cannot be before start date");
-            return;
-        }
 
         setEducationList((prev) => [
             ...prev,
